+++ backend-pegasus/api/routes/student/deletestudentjobpref.js
@@ -1,7 +1,11 @@
 const {env, sha1, mysql, mypool} = require('../../util')
 
 const deletestudentjobpref = (req, res) => {
-    const studentid = req.params.StudentID;
+    const studentid = req.body.StudentID;
+    const jobprefid = req.body.JobPreferenceID;
+    const industry = req.body.Industry;
+    const workexp = req.body.WorkExp;
+    const location = req.body.Location;
 
     mypool.getConnection( (error, connection) => {
         if(error) {
@@ -11,7 +15,10 @@ const deletestudentjobpref = (req, res) => {
         }
         else {
             if(studentid) {   
-                let queryString1 = `delete from pegasus.studentjobpref where studentid = "${studentid}"` ;         
+                let queryString1 = `delete from pegasus.studentjobpref where id = "${jobprefid}" and studentid = "${studentid}"` ;   
+                if(!studentid) {
+                    queryString1 = `delete from pegasus.studentjobpref where studentid = "${studentid}"` ;   
+                }       
                 connection.query(queryString1, (err, rows, fields) => {
                     if(err) {
                         res.status(500).json({ message: err })
